+++ mpi/compute_eqlbrmdistrfuncDF1.c
@@ -52,19 +52,19 @@ void compute_eqlbrmdistrfuncDF1(LV lv){
   Fluidnode* nodes;
 
   int ksi;
-  long BI, BJ, BK; //to identify the Sub grids
-  long cube_idx;
-  long starting_x, starting_y, starting_z, stopping_x, stopping_y, stopping_z;//To identify buffer zone
-  long li, lj, lk, node_idx;//local access point inside cube
+  int BI, BJ, BK; //to identify the Sub grids
+  int cube_idx;
+  int starting_x, starting_y, starting_z, stopping_x, stopping_y, stopping_z;//To identify buffer zone
+  int li, lj, lk, node_idx;//local access point inside cube
 
-  long  total_sub_grids, dim_x, dim_y, dim_z;
+  int  total_sub_grids, dim_x, dim_y, dim_z;
   dim_x = gv->fluid_grid->x_dim;
   dim_y = gv->fluid_grid->y_dim;
   dim_z = gv->fluid_grid->z_dim;
   int cube_size = gv->cube_size;
-  long num_cubes_x = gv->fluid_grid->num_cubes_x;
-  long num_cubes_y = gv->fluid_grid->num_cubes_y;
-  long num_cubes_z = gv->fluid_grid->num_cubes_z;
+  int num_cubes_x = gv->fluid_grid->num_cubes_x;
+  int num_cubes_y = gv->fluid_grid->num_cubes_y;
+  int num_cubes_z = gv->fluid_grid->num_cubes_z;
   total_sub_grids = (dim_x*dim_y*dim_z) / pow(cube_size, 3);
 
   int temp_mac_rank, my_rank;
